move check_password.js to common_functions
